# [ROCm] performance optimization by increasing the default max threads…