… per block
As observed during working on this fix (https://github.com/pytorch/pytorch/pull/130994), 128 threads per block seems quite low. This PR is to increase the default to improve the performance, and also slightly refactoring the code for better maintenance.

By increasing the default max threads per block from 128 to 256, I saw for `aten::index_select`,  its "CUDA total" time drop from 44.820ms to 33.608ms for by profiling below embedding script: 
```
input = torch.randint(low=0, high=16032, size=[size_of_input], device="cuda")
w = torch.randn([16032, size_of_column], device="cuda")

with profiler.profile(record_shapes=True) as prof:
    x = torch.nn.functional.embedding(input, w)

```
I did tested with the default from 256, 512, 1024 on MI210, and observed "CUDA total" time dropping even more as the number increases.
128 | 
256 | 33.608ms
512 | 31.001ms
1024 | 30.078ms

Fixes #ISSUE_NUMBER