@hongxiayang 2024-07-26
@eqy Please let me know whether you are comfortable for me to change the default to 256 for Nvidia case. We can change all the places of 128 to 256 in this pull request or next one.